# perf(python): Optimize sequence buffer writing via the c-api