--- a/BUILD
+++ b/BUILD
@@ -63,6 +63,7 @@ pyx_library(
     deps = [
         "//cpp/fury/util:fury_util",
         "//cpp/fury/type:fury_type",
+        "//cpp/fury/python:_pyfury",
         "@com_google_absl//absl/container:flat_hash_map",
     ],
 )