@penguin-wwy 2025-02-21
ci download bazel timeout, may retry directly?